* Separated rendering into its own module
