+++ fabric/modules/render/src/main/java/com/sigmundgranaas/forgero/render/texture/MinecraftResourceTextureProvider.java
@@ -1,4 +1,4 @@
-package com.sigmundgranaas.forgero.armor.client;
+package com.sigmundgranaas.forgero.render.texture;
 
 import com.sigmundgranaas.forgero.model.rendering.api.TextureProvider;
