+++ settings.gradle
@@ -33,3 +33,4 @@ include "fabric:modules:trade"
 include "fabric:modules:generator"
 include "fabric:modules:predicate"
 include "fabric:modules:armor"
+include "fabric:modules:render"
